ENT-3276: refactor test_managercli.py by modules
Refactor the big test_managercli.py, moving away from it:
- the tests for cli_command submodules for subscription-manager
  commands to own submodules
- the tests for functions in printing_utils to an own test

Card ID: ENT-3649
