+++ test/cli_command_test/test_facts.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from ..test_managercli import TestCliProxyCommand
+from subscription_manager import managercli
+
+
+class TestFactsCommand(TestCliProxyCommand):
+    command_class = managercli.FactsCommand
