+++ test/cli_command_test/test_plugins.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from ..test_managercli import TestCliCommand
+from subscription_manager import managercli
+
+
+class TestPluginsCommand(TestCliCommand):
+    command_class = managercli.PluginsCommand
