+++ test/cli_command_test/test_redeem.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from ..test_managercli import TestCliProxyCommand
+from subscription_manager import managercli
+
+
+class TestRedeemCommand(TestCliProxyCommand):
+    command_class = managercli.RedeemCommand
