+++ test/cli_command_test/test_refresh.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+from ..test_managercli import TestCliProxyCommand
+from subscription_manager import managercli
+
+
+class TestRefreshCommand(TestCliProxyCommand):
+    command_class = managercli.RefreshCommand
+
+    def test_force_option(self):
+        self.cc.main(["--force"])
