+++ test/cli_command_test/test_version.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from ..test_managercli import TestCliCommand
+from subscription_manager import managercli
+
+
+class TestVersionCommand(TestCliCommand):
+    command_class = managercli.VersionCommand
